+++ ChoGGi's Library/Code/CommonFunctions.lua
@@ -12,18 +12,6 @@ local FilterObjectsC = FilterObjectsC
 local StringFormat = string.format
 local TableRemove = table.remove
 
--- simplest entity object possible for hexgrids (it went from being laggy with 100 to usable, though that includes some use of local, so who knows)
-DefineClass.ChoGGi_HexSpot = {
-	__parents = {"CObject"},
-	entity = "GridTile"
-}
-DefineClass.ChoGGi_Vector = {
-	__parents = {"Vector"},
-}
-DefineClass.ChoGGi_Sphere = {
-	__parents = {"Sphere"},
-}
-
 -- backup orginal function for later use (checks if we already have a backup, or else problems)
 function ChoGGi.ComFuncs.SaveOrigFunc(class_or_func,func_name)
 	local ChoGGi = ChoGGi