+++ Mods ChoGGi/Loading Screen Example/metadata.lua
@@ -1,0 +1,17 @@
+return PlaceObj("ModDef", {
+	"title", "Loading Screen Example v0.1",
+	"version", 1,
+	"saved", 1539172800,
+	"image", "Preview.png",
+	"id", "ChoGGi_LoadingScreenExample",
+	"steam_id", "1535171613",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[Adds a single loading screen as an example to make your own pack.
+This will block the old ones, see Script.lua to add them back as well.
+
+Requested by... A friend.]],
+})