+++ Mods ChoGGi/RC Mechanic/metadata.lua
@@ -1,7 +1,7 @@
 return PlaceObj("ModDef", {
-  "title", "RC Mechanic v0.1",
-  "version", 1,
-  "saved", 1538481600,
+  "title", "RC Mechanic v0.2",
+  "version", 2,
+  "saved", 1539086400,
 	"image", "Preview.png",
   "tags", "Buildings",
   "id", "ChoGGi_RCMechanic",