+++ Mods ChoGGi/RC Miner/metadata.lua
@@ -1,7 +1,7 @@
 return PlaceObj("ModDef", {
-  "title", "RC Miner v1.3",
-  "version", 13,
-  "saved", 1538481600,
+  "title", "RC Miner v1.4",
+  "version", 14,
+  "saved", 1539086400,
 	"image", "Preview.png",
   "tags", "Buildings",
   "id", "ChoGGi_PortableMiner",