+++ Mods ChoGGi/Random Building Colour/metadata.lua
@@ -1,0 +1,16 @@
+return PlaceObj("ModDef", {
+	"title", "Random Building Colour v0.1",
+	"version", 1,
+	"saved", 1538913600,
+	"image", "Preview.png",
+	"id", "ChoGGi_RandomBuildingColour",
+	"steam_id", "1535187230",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[Changes the colours of objects to random colours when placed.
+
+Requested by McKaby]],
+})